+++ packages/apps/housing_loan_calculator/src/functions/calc-principal-equal-payment.ts
@@ -1,4 +1,4 @@
-import { seq } from '@noshiro/ts-utils';
+import { seq, uint32 } from '@noshiro/ts-utils';
 
 export const calcPrincipalEqualPayment = ({
   borrowingPeriodMonth: numPayments,
@@ -16,7 +16,7 @@ export const calcPrincipalEqualPayment = ({
 } => {
   const fixedPrincipalYenPerMonth = borrowingTotal / numPayments;
 
-  const borrowingBalanceYen = seq(numPayments + 1).map(
+  const borrowingBalanceYen = seq((numPayments + 1) as uint32).map(
     (i) => borrowingTotal - i * fixedPrincipalYenPerMonth
   );
